sleep(POLL_TIME) before checking state change in enable, disable functions
Recently, The bluetooth adpater introduced a statemachine to handle state
change. The enable and disable method post a message to the state machine.
The statemachine starts the switch in a different thread. There might be a
little bit delay of the real state change.
We change the test cases to wait a little bit before checking state change.
bug5092306

Change-Id: I14eceda9a3c9b3ad31cac3e6418fb93afdb3fc03