ActionBar tests.
Uses a new framework for the action bar tests
that has an activity that takes its own snapshot.
These are then compared or saved as necessary.
Tests are run in both portrait and landscape
and are also run with uiOptions="SplitActionBarWhenNarrow".

To fix orientation errors, we now retry the specific
test on that specific theme whenever a
failure case is detected.

Change-Id: I6b9d0e7b83ceba2b1e9cfc211d1dec707e93a1c3