Allow Tests to Require Some Features
Bug 5104638

Add the ability to specify what features are required for a test.
If the device doesn't have all of the required features then the
test will not appear in the menu. Specify the required features
by specifying a meta-data entry in the Android manifest.

This does create a possibility where a device doesn't report the
right features and doesn't see a test that it should. However,
they should review the features in the feature summary activity
I hope...

...also remove some unused resources!

Change-Id: I8233510c586f3b2547ac40cd0502b9c6b7d03484