Renaming DelayedCheck to PollingCheck
This should not change functionality in any way.

Better name, also moving to package android.cts.util when I'm
already there doing renaming.

Want to modify it to check before the first wait as well, but will
wait for this check to be commited first so we can better track
any new test failures.

Change-Id: If6281dd59cb3b183e39c4b9c4f89364e69811751