Make zergrush test more reliable.
Don't use /proc/net/netlink to determine if vold is alive. It's possible
for a previous test to cause vold not to be listening on /proc/net/netlink.
Instead, scan /proc and determine the PID from there.

Get rid of the IllegalStateException fail(). This was confusing and
hid the actual warning message we want people to see.

Change-Id: I78791fd6e8a539d980351c6e1eb51cc9dab55fb5