Fixed a broken rtsp test case
The original cts test case had an incorrect requirement that onVideoSizeChanged() callback must be received
before streaming or playback can be started. For http live streaming or RTSP streaming, the video size can
only be received sometimes after the video streaming has started. As a result, the failed test case never
triggered the streaming of the rtsp videos.

Change-Id: Icca4dd216b8b99993882a7bf11b80503909f47c3
related-to-bug: 5790219