Fix CTS WebViewTest#testFindAddress
 Bug: 5866557

The address parser now recognises 929 as an invalid zipcode prefix, and
rejects addresses from that zipcode. Our test case happened to use that
zipcode, and was incorrectly expecting an invalid address to be accepted.

I've added a test case to expect failure on this (invalid) zipcode prefix,
and adjusted the old cases to use a valid zipcode prefix instead.

Change-Id: I3c5a75734c22674e1828dbf33f3a94dda758c238