Refactor, fix and expand CTS test for playlists
The CTS test for playlists wrongly assumed that when reordering items
in a playlist, the specified positions were tplay_order values, not
the actual position in the list. The test just happened to succeed
because of another bug in the media provider.
This change fixes these tests, adds more tests, and refactors the
code a little to make it easier to read and expand in the future.

b/6299084

Change-Id: I010714074e4137b17d28e00b007b0edc1e0ea27f