Fix AnimationDrawableTest
A previous change fixed the logic about assuming an AnimationDrawable()
was not running after a call to setVisible(true, true). This CL fixes
another such issue, where we assume that we should still be on the same
frame of the animation because the animation is not running (which is
incorrect).

Change-Id: I7af09bca9ff005f88b4426b7ff3d66fe1874891d