Updated SpannableStringBuilderSpanTest for correct MARK/POINT
Bug 6343982

Updated the test to introduce the different behavior of POINT and
MARK span boundaries when some text is replaced by an other text
(as opposed to a deletion or insertion).

Note that the original implementation of SpannableStringBuilder
(with an added space character followed by three calls to change() hack)
does not pass this test.

Change-Id: I62569a9f73d100276300dae9119a05bd4f61eb10