Change the test case about SCREENLAYOUT_LONG  depending whether the navigation bar is displayed or not.
For example
  1. The device has WXGA (1280x800) display.
  2. The navigation bar is displayed
  3. The height of navigation bar is 64pt.

At landscape mode, the height is 736 (=800-64) and  the width is 1280,
then the device is considered as wider than WVGA.
(i.e. SCREENLAYOUT_LONG_YES)

At portrait mode, the height is 1216 (=1280-64) and the width is 800,
then the device is not considered as wider than WVGA.
(i.e. SCREENLAYOUT_LONG_NO)

I modified to check the navigation bar is displayed or not.