Improve OpenGL test reliability
These tests were launching activities, setting a view, and then sleeping
for 1 second. Send a couple extras via an intent to the test activity
so it creates the content view properly in one shot instead. This should
remove the need for sleeping and fix the occasional NPEs about view
being null.

Bug 6485813

Change-Id: I861cf74bdae3c0b46cba7d58c1218acf1d6297a5