New StaticLayout line break tests
These test exercise the generate() method of StaticLayout which splits
the provided text into lines.

The tests annotated with a "Bug" comment, although they have been changed
to pass should indeed be fixed.

The use of Spans bypasses the custom TextPaint passed to the StaticLayout
and MeasuredText#addStyleRun has to be modified to pass the tests that
add Spans to the text. These tests will be disabled before the commit
to make the test green.

Added a test with 2 spans, with all the possible span combinations.

Change-Id: I83280c95fff844a101758f785e838e32b80b114e