Fix pattern issue in MonkeyTest
The pattern is designed to match the output of logcat without
format.
But, there are 8 formats defined in /system/core/liblog/logprint.c.
If logcat is set with any format, the test cases will fail.
Since 'executeAdbCommand' ensures the tag filter of 'MonkeyActivity:I',
check if the output contains MONKEY or HUMAN should work as expected.

Change-Id: Ibcf18407bf6dee0d405ca90c5d597cff940a7201