Fix CTS test failure
CTS test case android.telephony.cts.TelephonyManagerTest#testGetDeviceId
was failed on CDMA+LTE device.
When phone is in CDMA+LTE, Android considers it to be PHONE_TYPE_CDMA
and in turn TelephonyManager.getDeviceId() will return MEID as device
ID. However, LTE is a 3GPP protocol, and reasonable to use IMEI for
device ID when phone is in LTE mode.
So, the fix is: added a check for LTE mode to use IMEI instead of
MEID.

Change-Id: I0b42555a9aafb79093218c4fc4b4ca0c06d2fb59