DO NOT MERGE Filter out the overdue location cases in GPS test
Bug 7109051

This was merged into jb-mr1 already.

Occasionally, passive listeners receive a overdue location.
The overdue location comes from previous test round, and it will
confuse calculation of time interval. Finally, the confused interval
is mistaken for a FAIL even though GPS works normally. For this reason,
we have to filter out the overdue location at beginning of test round.
We log that the overdue location was ignored.

Change-Id: Ia05c973cb0010e1df1bf2c07a85fdae722ec5c94