Delete a bad CTS test.
This test calls WakeLock.release() without previously
having acquired the WakeLock (because it does not have
permissions to do so).  The release() turns into a no-op
on the client side because the WakeLock is not held.
Consequently the system will not check permissions
(it doesn't have to) so no SecurityException is thrown.

That said, the system does check permissions on release()
but we can't test that through the API.

Bug: 6950370
Change-Id: I4a94b33b223e8010c285b756ae7d6bbbb4f308a1