Fix the failing cts test
It seems like testClearSslPreferences is failing (was passing fine
locally). The test verifies the url in the callback by comparing it
to the url in the loadUrl() call, however, this seems not a good
comparison since there is a race between loading favicon and the
page url. Change it to check for the page title.

Change-Id: I178ef17c7cdb8126b89d82b95f203fdbe7b9c82e