Refactor CtsTestServer for WebViewTest.testLoadDataWithBaseURL
The test web server only keeps track of the last requested URL. The
order in which requests are received by the server is highly
implementation dependant and non-deterministic.

Refactor the test server to keep track of all requests, and update
assertions in tests to ensure that the request was made at some point.
This should eliminate flakiness.

Bug: 7907196
Change-Id: Iaf70383f9a136c25028d9ec405301ccf86ecbe53