Improving CTS test stability.
The CTS tests for accessibility text traversal rely on exact content being
laid out on every screen so we can verify that exactly one page was advanced
while traversing. The tests were relying on attributes taken from the theme
which can be changed by vendors. Also the text size on the platform was able
to affect the test. Now the layout is independent of external state.

Change-Id: If4ec9db8a1fc2180c7888eb56d7627444e44002c