Update WebHistoryItemTest to be more deterministic.
Favicon loading is always on with the Chromium WebView, so when
testing favicon behavior make sure we've turned it on for the
Classic WebView.

Additionally, ensure that we've received an a favicon before taken
the history snapshot so that we've always got something to compare.

Finally, compare the content of the bitmaps rather than the pointers
(aside - due to the issues outlined above, this was always comparing
null with null in the Classic case)

Bug: 9121594
Change-Id: Ic141d5c352bf9327087c0b410aab499b06dd67c6