Implementing Jank Test for UI Scrolling.
The UI scrolling tests share a lot with the OpenGL tests so common code
has been pulled up into a new super class. The ScrollingActivity has
also changed to allow the number of elements to be given through the
intent, this should not affect the existing ScrollTest as the default
value is the same as it was before.

Change-Id: I1ca58f535ac81bc0e9032934758dc5a49dbac0fb