Assert that /dev/random and urandom are world-readable/writable.
This CL adds tests into CTS to assert that the Linux RNG character
devices /dev/random and /dev/urandom are world-readable and
world-writable. This is needed mainly to ensure that apps can use
/dev/urandom (and /dev/random) as sources of randomness.

Some of the existing CTS tests are already implicitly checking that
/dev/urandom is world-readable. However, it is time to stregthen the
contract and assert the world-readability/writability explicitly.

Bug: 10394220
Change-Id: I772c0472f2367a5da14ef1613cf187c230f55d51