Fix flaky EmojiTest.testEmojiGlyph.
The WebView is rendering the emoji correctly but capturePicture
sometimes captures an earlier frame that hasn't yet been rendered with
the new content. Determining when the new picture is ready reliably is
very difficult (and not something the Chromium-powered WebView's
rendering pipeline really supports), but sleeping for a short time is
sufficient as the page to be rendered is very simple and will have
little variance in rendering time.

Bug: 10946703
Change-Id: Ib08511f33205808ea7c19f8ae81f06c36c6780fe