Test decoding a short video stream.
Add CTS test that decodes one of the H.264/AVC resource videos but truncates
the stream by signaling end-of-stream shortly after the start of the input. In
this situation the decoder should output decoded frames up to the end of stream
(assuming the input bitstream to the truncation point is complete).

Bug: 11433909
Change-Id: I4de055d698a4af517d0dbbe81ff10724186ea515