Initial SELinux CTS test based on current policy file.
This test takes as input an xml file based on the parsing of our SELinux
policy.conf file.  An example script for generating such an xml file from the
AOSP intermediate policy.conf file may be found at
platform/packages/experimental/SELinux/CTS/ which is described in more detail
in https://android-review.googlesource.com/#/c/83292/

The xml file contains entries for every valid allow and
neverallow rule present in the input policy file and represents them as groups
of values to be checked as part of selinux security contexts sent to the selinux
library selinux_check_access() function.  This function indicates whether
the source security context may use the indicated permission for a particular
object class with the target security context. This test is tied intimately to
the particular policy.conf file used to generate selinux_policy.xml. This is
initially based on a core set of neverallow rules which should be true of all
android devices.

Change-Id: I35a28fc010607ec55d9e39649343bca02356e1ed