Fix for failing test of testEOS because checksum takes too long
While low latency port reconfiguration is on (the same as smooth
streaming but for local playback) allocated buffers has maximum
supported capacity in our case it is Full HD.
Effective buffer size is equal to HxWx3/2 but in this test capacity
was used for checksum check. So instead of check of buffer ~300k
test was checking buffer of ~3M size what makes the test to take
too long and test failed due to timeout.

In test used space of the buffer should be used.

Change-Id: Ie4854eef1d26737bfeade298ebb80f91891c8b6d
(cherry picked from commit 923bb77624d140dcf49c0eedf0c2226458e8cdd4)