Fix a bug regarding layout-overlapped buttons in IntentDrivenTestActivity Class
If we test "Alarm and timers Tests" in Low-Density Device such as mdpi and ldpi,
Thoes two buttons which are "Set Alarm" and "Verify" are put under "Pass" and "Fail" buttons. we cannot press "Set Alarm" and "Verify" button.

I added a ScrollView as a root element to avoid that. following a Stuart Scott's advice.

Signed-off-by: Sangkyeong Jeong <sangkyeong.jeong@lge.com>

Change-Id: Ie4bf9a3ea6f7cc7a05fe4801c83fc62883169b5e