Verify the failed test case of CTS for WebView
Device resolution: 800*1280
Android version: 4.2

When we run the CTS on device in portrait mode, the test case (android.webkit.cts.WebViewTest#testSetInitialScale) always fail, instead, this test case passes in landscape mode. We assume that the cause of this issue is the device resolution is higher than the resolution of WebView page in portrait mode. Also, the same issue can be reproduced on Nexus 7 (Android 4.2) with the same resolution as our device.

This test case uses a virtual site which is 1000*1000, when the height of tablet screen is greater than 1000, the layout of WebView will not scale the page size, however, when the height is smaller than 1000, the scrollbar will disappear and the site can be scaled to 50% successfully. If we change the default setting from 1000*1000px to 2000*2000px, then this test case (setInitialScale) will pass. Moreover, we notice that the setting of CTS for Android 4.4 has changed to 2000*2000px.

Please help to check this issue, Google can try to run this test case on Nexus 7 (Android 4.2) in portrait mode, the test case fails every time. 

Change-Id: Ide7db52ea24e8c1d48126bea57af29528115bea9
Signed-off-by: endy <endy.wei@ecs.com.tw>