Extend WebP encoder test for PNG(Alpha) image.
There was a bug in the WebP encoder while encoding PNG images with
Alpha. Fixed that bug (Refer:
https://googleplex-android-review.git.corp.google.com/410613).
Extended the WebP encoder test to test the encoding of PNG image with
Alpha channel.
This cts test change have pre-requisite changes and should be submitted once those two changes are submitted.
https://googleplex-android-review.git.corp.google.com/410613
https://googleplex-android-review.git.corp.google.com/412038

Change-Id: I60fbc903c0c2346d1274c06d54021daba486e1bf