Fix mac SDK builds.
libBionicTests isn't built for Mac hosts, so we can't
build bionic cts tests either. Even though CTS tests run
on the target (not the host) bionic-unit-tests-cts_list
can't be built, and is used to generate the test
description XML.

Note: This shows up as a build breakage on
mirror-aosp-master but not master because the build bot
doesn't use checkbuild on that target.

Change-Id: I23b980160945f071264cfbb085a796416d78afb5