Add Vivante exceptions for FileSystemPermissionTest
/dev/galcore or /dev/graphics/galcore is used by all GL applications
and HWC for accessing Vivante GPU and CPC engine. All 3D and 2D contexts
are protected from each other with GMMU (GPU Memory Management Unit).
The interface is secured by maintaining a whitelist of registers that
are allowed to be accessed from usermode. Vivante had carefully reviewed
/dev/galcore or /dev/graphics/galcore interfaces in terms of security
concern, and we are sure the device meets Android security requirements.

Change-Id: I059ffc673762678af3ad6abb04029454f879b6c7
Signed-off-by: Zongzong Yan <zongzong.yan@vivantecorp.com>