Fix a concurrency bug in OpenSSLHeartbleedTest.
The test was not guaranteed to observe all of the data transmitted by
the client in the scenarios where the client failed. As a result, the
test was occasionally failing (instead of passing) when the client
terminated the TLS/SSL handshake with the fatal alert
unexpected_message.

This CL fixes the test by ensuring that the test observes all of the
data exchanged between the client and the server.

Bug: 13906893

(cherry picked from commit 352696d5699051ac2f92af8b0a1156ade0fc59a6)

Change-Id: I5091ae7f5e12bdb6fe76a9f0f44618e0e55042c5