Fix CookieTest CTS failures for No Webview Devices
Seems like the problem was a copying error.
All the other tests check if getActivity().getWebView() is null
in setup but CookieTest and CookieSyncManagerTest uses
NullWebViewUtils.isWebViewAvailable instead.
I think the reason this is wrong is beacuse onCreate sets up
NullWebViewUtils and it isn't called until we call getActivity()
for the first time.

Bug: 15326886
Change-Id: I04f8344e232fdd8a0fd1260cc8d915909109d851