Fix some bad tests for ArrowKeyMovementMethod.
It turns out that the tests for the DPAD UP, LEFT, and RIGHT tests
all passed because they were correctly including modifier metastates
in the KeyEvents they generated but the DPAD DOWN test was styled a
little differently and didn't so it broke.

These tests were also passing ACTION_UP instead of ACTION_DOWN to
onKeyDown() although that was ignored by the function.

The test for DPAD DOWN might have worked as written when the default
keyboard behavior was to use toggled modifiers but the virtual keyboard
device actually expects chorded modifiers which is what caused the
problem.

Bug: 15460082
Change-Id: Ibc1e6c6145062e522cb8f1887575c4885e5fa29a