Require seccomp if we're running a 3.8 or greater kernel.
We'll be doing future work on seccomp, and third party applications,
such as Chrome on Android, need seccomp for sandboxing.

Seccomp can be enabled in the kernel by setting CONFIG_SECCOMP_FILTER=y

For kernels prior to 3.8, we don't require seccomp, as it would
require backporting work for kernel developers.

Bug: 16409560
Change-Id: I44b77510a46b2855a34e37b12ad56248d2f89657