Fix issue with android.widget.cts.TextViewTest#testGetFadingEdgeStrength are failing on ro.sf.lcd_density=120
Run cts android.widget.cts.TextViewTest#testGetFadingEdgeStrength would be fail on the device with
(low density/120 setting in Android Display with 1920x1080/17.3 inch panel).
(low density/240 setting in Android Display with 3840x2160/28 inch panel).

The length of text (named “textViewLeft” in code) for CTS test is smaller than the visible area of app’s content in screen.
After increasing size of “textViewLeft” which exceeds the width of the view, it can pass the verification.
Expanded it to fix the problem.

Change-Id: I78c72f9a92023fefe8f740853cc339bea56c83c4

Conflicts:
	tests/res/values/strings.xml