Fix SensorIntegrationTests to use supported sensor rates.
b/16896822

Sensor Cts tests did add suppor to error out if events did not arrive in a given window.
Because UI/GAME/NORMAL translation values are not posted in CDD, it is not possible to calculate
the time events would take to arrive.

This change fixes the tests that were affected after that change in SensorCtsHelper.

Change-Id: Idde9bff416c77045701d1a9c16fb28a9a21ebc55