Exception in TestAccessHistory tearDown
CTS test will report a exception in TestAccessHistory
method tearDown, when having the AOSP Browser installed.
The reason for this is that in setup the bookmarks are saved
as a backup and in teardown they are restored to the data base.
But the BrowserProvider in AOSP Browser does not delete the
bookmarks, only mark them as deleted. So in tearDown when the
bookmarks is restored there will be a exception in SQLite
because the ID exist.
Fix: Do not include the ID when backing up the bookmark

Change-Id: Iba1031958bffbfcccd03ae792ced2d7ffde9b5bf
(cherry picked from commit ed80c0eab9f388a3291f2005b2242a956dc58f7e)