SELinuxHostTest:  Check that each executable has the expected domain.
The current SELinux domain tests only check that each domain has
the expected executable, but not vice versa, so it is possible
for a service to run in a completely different domain and still
pass the test.  Validate that each running process with one of these
executables has the expected domain as well.

Change-Id: I927da0736137db9a807f92d09e70940f235a7b8e
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>