Fix cts coverage report inclusion of incorrect classes
CTS coverage reporting uses dexdeps to determine which classes
and methods are used by a test app. Using LOCAL_STATIC_JAVA_LIBRARIES
to link dependent java libraries, causes the dependent libraries' dependencies
to be directly included in the classes.dex and therefore in the output
from dexdeps. This causes to the coverage tool to report coverage of
methods and classes that are not actually access by the test applications.

Switching to LOCAL_JAVA_LIBRARIES prevents this and does not appear to
break these tests.

bug: 21448392
Change-Id: Id031c71e409450225a9287e796a0fda53203bd8b