Add: test case for SettingInjectorSerivce
A test case for android.location.SettingInjectorSerivce. This is a
standalone test case that only test the functionality of
SettingInjectorSerivce. Passing this test do not guarantee
interoperability with Android OS and other component. For an end-to-end
test please refer to the GTS UI automator test case in
com.google.android.xts.location.InjectedSettingTestCase.

Change-Id: I21ead86006dd08e6f349536d8d4d76bdc14bd653