Remove cell / wifi manipulation from ApacheHttpClientTest
The testcase was manipulating the Wifi / Cell networks,
presumably in an effort to make data pass over each type of
network.

The test was sending / receiving data using the loopback
interface so it doesn't make much sense.

testExecute_withMobile would fail if the device did not
have a SIM / mobile data connection *and* if the wifi
network was up when the test ran; it would fail waiting
for the mobile data network to enter the CONNECTED state.
If the wifi connection was down at the start it would
assume the mobile data connection was up already and
proceed with the test (which passed, because it
didn't actually use the network).

This change chops out parts of this test to leave just
HttpClient-related code.

Bug: 21874093
Change-Id: I8248b840f58f939661eff2345f70a9e786e593f3